Make another module with just one function
Because alexey likes files
