+++ dmutils/apiclient/__init__.py
@@ -1,5 +1,5 @@
 from .errors import APIError, HTTPError, InvalidResponse
 from .errors import REQUEST_ERROR_STATUS_CODE, REQUEST_ERROR_MESSAGE
 
-from .data import DataAPIClient, user_has_role
+from .data import DataAPIClient
 from .search import SearchAPIClient
