+++ dmutils/apiclient/data.py
@@ -293,10 +293,3 @@ class DataAPIClient(BaseAPIClient):
                     "update_reason": reason,
                 },
             })
-
-
-def user_has_role(user, role):
-    try:
-        return user['users']['role'] == role
-    except (KeyError, TypeError):
-        return False
