+++ tests/test_apiclient.py
@@ -10,7 +10,6 @@ import mock
 from dmutils.apiclient.base import BaseAPIClient
 from dmutils.apiclient import SearchAPIClient, DataAPIClient
 from dmutils.apiclient import APIError, HTTPError, InvalidResponse
-from dmutils.apiclient import user_has_role
 from dmutils.apiclient.errors import REQUEST_ERROR_STATUS_CODE
 from dmutils.apiclient.errors import REQUEST_ERROR_MESSAGE
 
@@ -978,19 +977,3 @@ class TestDataApiClient(object):
                 'updated_by': 'user'
             }
         }
-
-
-def test_user_has_role():
-    assert user_has_role({'users': {'role': 'admin'}}, 'admin')
-
-
-def test_user_has_role_returns_false_on_invalid_json():
-    assert not user_has_role({'in': 'valid'}, 'admin')
-
-
-def test_user_has_role_returns_false_on_none():
-    assert not user_has_role(None, 'admin')
-
-
-def test_user_has_role_returns_false_on_non_matching_role():
-    assert not user_has_role({'users': {'role': 'admin'}}, 'supplier')
